Add S_SoundSampleIsPlaying();
Add S_SoundSetPitch();
Continuity fixes to CD.H (Gamewad_header struct)
Add more methods from CODE.WAD to SPECIFIC.C/.H
Correct S_PlayFMV param data types.
HEALTH.C condition bug fix.

+++ SPEC_PSXPC/FMV.C
@@ -2,7 +2,7 @@
 
 #include "CONTROL.H"
 
-long S_PlayFMV(long sequence, long flag)//*, (<)5E7A0
+int S_PlayFMV(int sequence, int flag)//*, (<)5E7A0
 {
 	//DISPENV disp[2];
 	//s1 = fmv_index
